renamve hook files
