+++ playground/src/components/module_config/index.tsx
@@ -2,7 +2,7 @@ import { Dispatch, SetStateAction, useEffect, useState } from "react";
 import validator from "@rjsf/validator-ajv8";
 import { getDefaultFormState } from "@rjsf/utils";
 
-import { useConfigSchema } from "@/hooks/use_config_schema";
+import { useConfigSchema } from "@/hooks/config_schema";
 import { ModuleMeta } from "@/model/health_response";
 
 import DefaultSchemaFormModuleConfig from "./default_schema_form";
